contributors/adobe/submitted/exclusions/wrap-margin-001.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>CSS Exclusions Test: wrap-flow property set to start</title>
        <link rel="author" title="Jacob Goldstein" href="mailto:jacobg@adobe.com">
        <link rel="help" href="http://dev.w3.org/csswg/css3-exclusions/#wrap-flow-property">
        <meta name="flags" content="ahem dom">
        <meta name="assert" content="">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="resources/helper.js"></script>
        <style>
            * {
                margin: 0px;
                padding: 0px;
            }
            .exclusion {
                -webkit-wrap-flow: start;
                -webkit-wrap-margin: 12px;
                position: absolute;
                top: 40px;
                left: 40px;
                width: 60px;
                height: 25px;
                background: blue;
            }    
            #content {
                width: 400px;
                line-height: 12px;
                font: 12px Ahem;
            }
        </style>
    </head>
    <body>
        <div class="exclusion"></div>

        <div id="content">
            <span id="linesAbove">ethical<BR>beer<BR>craft<BR></span>
            <span id="lineRight1">foo</span>
            <span id="lineRight2">sed</span>
            <span id="linesBelow">Vegan<BR>cliche<BR>retro</span>
        </div>
        
        <div id="log"></div>
        
        <script type="text/javascript">       
            test(function() {assert_equals(checkLinePos("lineRight1",36,"top"), true)}, "Verify top of the first line to the left of the exclusion is positioned correctly");
            test(function() {assert_equals(checkLinePos("lineRight1",110,"left"), true)}, "Verify top of the first line to the left of the exclusion is positioned correctly");
            test(function() {assert_equals(checkLinePos("lineRight2",48,"top"), true)}, "Verify top of the second line to the left of the exclusion is positioned correctly");
            test(function() {assert_equals(checkLinePos("lineRight2",110,"left"), true)}, "Verify right of the second line to the left of the exclusion is positioned correctly");
            test(function() {assert_equals(checkLinePos("linesBelow",75,"top"), true)}, "Verify top of the line below the exclusion is positioned correctly");
        </script>
    </body>
</html>
